# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Fingarden API Docs",
    "version": "1.0.0",
    "title": "Fingarden"
  },
  "basePath": "/",
  "paths": {
...
    "/subscribes": {
      "get": {
        "tags": ["subscribes"],
        "summary": "Get array of subscribers",
        "description": "",
        "operationId": "getSubscribers",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "admin-token or user-token"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SubscribersResponse"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["subscribes"],
        "summary": "Create a subscriber",
        "description": "",
        "operationId": "createSubscribers",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/SubscriberDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/SubscribersResponse"
              }
            }
          },
          "409": {
            "description": "Conflict"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/checkouts/all": {
      "get": {
        "tags": ["checkoutAll"],
        "summary": "Get array of checkouts",
        "description": "",
        "operationId": "getCheckoutsAll",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "addressId",
            "in": "query",
            "description": "Filtering checkouts by addressId",
            "required": false,
            "type": "string"
          },
          {
            "name": "basketId",
            "in": "query",
            "description": "Filtering checkouts by basketId",
            "required": false,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "query",
            "description": "Filtering checkouts by userId",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sorting checkouts by property",
            "default": "basket",
            "required": false,
            "type": "string"
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Ordering results. ASC | DESC",
            "default": "DESC",
            "required": false,
            "type": "string"
          },
          {
            "type": "string",
            "in": "query",
            "name": "offset",
            "required": false,
            "default": 0,
            "description": "qty rows to skip for pagination"
          },
          {
            "type": "string",
            "in": "query",
            "name": "limit",
            "required": false,
            "default": 10,
            "description": "qty rows to take for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CheckoutResponse"
              }
            }
          }
        }
      }
    },
    "/checkouts": {
      "get": {
        "tags": ["checkout"],
        "summary": "Get array of checkouts",
        "description": "",
        "operationId": "getCheckouts",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "addressId",
            "in": "query",
            "description": "Filtering checkouts by addressId",
            "required": false,
            "type": "string"
          },
          {
            "name": "basketId",
            "in": "query",
            "description": "Filtering checkouts by basketId",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sorting checkouts by property",
            "default": "basket",
            "required": false,
            "type": "string"
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Ordering results. ASC | DESC",
            "default": "DESC",
            "required": false,
            "type": "string"
          },
          {
            "type": "string",
            "in": "query",
            "name": "offset",
            "required": false,
            "default": 0,
            "description": "qty rows to skip for pagination"
          },
          {
            "type": "string",
            "in": "query",
            "name": "limit",
            "required": false,
            "default": 10,
            "description": "qty rows to take for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CheckoutResponse"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["checkout"],
        "summary": "Create a checkout",
        "description": "",
        "operationId": "createCheckout",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CheckoutDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful created",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Checkout request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/checkouts/{checkoutId}": {
      "get": {
        "tags": ["checkout"],
        "summary": "Find checkout by id",
        "operationId": "findCheckoutById",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "checkoutId",
            "in": "path",
            "description": "ID of checkout to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Checkout"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Checkout not found"
          }
        }
      },
      "put": {
        "tags": ["checkout"],
        "summary": "Update the checkout",
        "description": "",
        "operationId": "updateCheckout",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "checkoutId",
            "in": "path",
            "description": "ID of checkout to return",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "totalAmount": {
                  "type": "number"
                },
                "userId": {
                  "type": "string"
                },
                "comment": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful updated",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CheckoutWithoutJoins"
            }
          },
          "400": {
            "description": "Invalid Checkout request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Checkout not found"
          }
        }
      },
      "delete": {
        "tags": ["checkout"],
        "summary": "Delete the checkout",
        "description": "",
        "operationId": "deleteCheckout",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "checkoutId",
            "in": "path",
            "description": "ID of checkout to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful deleted",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "comment": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Checkout not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/api/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "Brand": {
      "required": ["name", "url"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "showOnMain": {
          "type": "boolean",
          "default": false,
          "description": "value true can be post only with admin token"
        }
      },
      "xml": {
        "name": "Brand"
      }
    },
...
    "Category": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "parent": {
          "type": "object",
          "$ref": "#/definitions/Category"
        },
        "children": {
          "type": "object",
          "$ref": "#/definitions/Category"
        },
        "parameters": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Parameter"
          }
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "Color": {
      "required": ["name", "url", "code"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/ProductWithoutJoins"
          },
          "readOnly": true
        },
        "url": {
          "type": "string"
        },
        "code": {
          "type": "string"
        }
      }
    },
...
    "Parameter": {
      "required": ["name"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Rating": {
      "properties": {
        "1": {
          "type": "number"
        },
        "2": {
          "type": "number"
        },
        "3": {
          "type": "number"
        },
        "4": {
          "type": "number"
        },
        "5": {
          "type": "number"
        },
        "avg": {
          "type": "number"
        }
      }
    },
    "ParameterProduct": {
      "properties": {
        "productId": {
          "type": "string"
        },
        "parameterId": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "parameter": {
          "$ref": "#/definitions/Parameter"
        }
      }
    },
...
    "Product": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "shortDesc": {
          "type": "string"
        },
        "keywords": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "category": {
          "type": "object",
          "$ref": "#/definitions/Category"
        },
        "brand": {
          "type": "object",
          "$ref": "#/definitions/Brand"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/TagWithoutJoins"
          }
        },
        "sizes": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/SizeWithoutJoins"
          }
        },
        "url": {
          "type": "string"
        },
        "rating": {
          "$ref": "#/definitions/Rating"
        },
        "reviews": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewWithoutJoins"
          }
        },
        "questions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Question"
          }
        },
        "parameterProducts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ParameterProduct"
          }
        },
        "productVariants": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductVariant"
          }
        }
      }
    },
...
    "ProductWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "url": {
          "type": "string"
        }
      }
    },
    "ProductVariant": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "price": {
          "type": "integer",
          "minimum": 1
        },
        "oldPrice": {
          "type": "integer"
        },
        "wholeSalePrice": {
          "type": "integer"
        },
        "available": {
          "type": "boolean"
        },
        "images": {
          "type": "string"
        },
        "color": {
          "type": "object",
          "$ref": "#/definitions/Color"
        }
      }
    },
...
    "TagWithoutJoins": {
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "SizeWithoutJoins": {
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "AddressWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "userId": {
          "type": "string"
        },
        "receiverName": {
          "type": "string"
        },
        "receiverPhone": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "roomOrOffice": {
          "type": "string"
        },
        "door": {
          "type": "string"
        },
        "floor": {
          "type": "string"
        },
        "rignBell": {
          "type": "string"
        },
        "zipCode": {
          "type": "string"
        }
      }
    },
...
    "Basket": {
      "required": ["orderProducts"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "userId": {
          "type": "string"
        },
        "orderProducts": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/OrderProduct"
          }
        },
        "totalAmount": {
          "type": "number"
        },
        "checkout": {
          "type": "object",
          "$ref": "#/definitions/CheckoutWithoutJoins"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "status": {
          "type": "string",
          "default": "New"
        }
      }
    },
...
    "Checkout": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "paymentId": {
          "type": "string"
        },
        "address": {
          "type": "object",
          "$ref": "#/definitions/AddressWithoutJoins"
        },
        "basket": {
          "type": "object",
          "$ref": "#/definitions/Basket"
        },
        "comment": {
          "type": "string"
        },
        "leaveNearDoor": {
          "type": "boolean"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "status": {
          "type": "number",
          "enum": ["New", "InDelivery", "Completed"]
        }
      }
    },
    "CheckoutResponse": {
      "properties": {
        "rows": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Checkout"
          }
        },
        "length": {
          "type": "number"
        }
      }
    },
    "CheckoutWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "paymentId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        },
        "leaveNearDoor": {
          "type": "boolean"
        }
      }
    },
    "CheckoutDTO": {
      "required": ["address", "basket"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "paymentId": {
          "type": "string",
          "description": "Payment ID"
        },
        "address": {
          "type": "string",
          "description": "Address ID"
        },
        "basket": {
          "type": "string",
          "description": "Basket ID"
        },
        "comment": {
          "type": "string"
        },
        "leaveNearDoor": {
          "type": "boolean"
        },
        "totalAmount": {
          "type": "number",
          "description": "Total payment"
        }
      }
    },
    "OrderProduct": {
      "required": ["product", "inBasket"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "product": {
          "type": "object",
          "$ref": "#/definitions/Product"
        },
        "productVariant": {
          "type": "object",
          "$ref": "#/definitions/ProductVariant"
        },
        "qty": {
          "type": "integer",
          "minimum": 1
        },
        "productSize": {
          "type": "string"
        },
        "productPrice": {
          "type": "integer",
          "readOnly": true
        },
        "inBasket": {
          "type": "object",
          "$ref": "#/definitions/Basket"
        }
      }
    },
...
    "Review": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "rating": {
          "type": "integer",
          "minimum": 1,
          "maximum": 10
        },
        "text": {
          "type": "string"
        },
        "images": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "showOnMain": {
          "type": "boolean",
          "default": false
        },
        "product": {
          "type": "object",
          "$ref": "#/definitions/Product"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        },
        "reactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewReaction"
          }
        }
      }
    },
    "Question": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "text": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "product": {
          "type": "object",
          "$ref": "#/definitions/Product"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/QuestionComment"
          }
        },
        "reactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReactionQuestion"
          }
        }
      }
    },
...
    "ReviewWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "rating": {
          "type": "integer",
          "minimum": 1,
          "maximum": 10
        },
        "text": {
          "type": "string"
        },
        "images": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "showOnMain": {
          "type": "boolean",
          "default": false
        },
        "productId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        },
        "reactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewReaction"
          }
        }
      }
    },
    "SubscribersResponse": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "SubscriberDTO": {
      "type": "object",
      "required": ["email"],
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
...
    "ReviewReaction": {
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "reviewId": {
          "type": "string"
        },
        "reaction": {
          "type": "string",
          "default": "Like"
        }
      }
    },
    "ReactionQuestion": {
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "questionId": {
          "type": "string"
        },
        "reaction": {
          "type": "string",
          "default": "Like"
        }
      }
    },
    "User": {
      "required": ["firstName", "lastName", "email", "password"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "isVerified": {
          "type": "boolean",
          "default": false
        },
        "role": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "Comment": {
      "properties": {
        "id": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "review": {
          "type": "object",
          "$ref": "#/definitions/Review"
        },
        "text": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "reactions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/CommentReaction"
          }
        }
      }
    },
    "QuestionComment": {
      "properties": {
        "id": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "question": {
          "type": "object",
          "$ref": "#/definitions/Question"
        },
        "text": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "reactions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/CommentReaction"
          }
        }
      }
    },
    "CommentReaction": {
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "commentId": {
          "type": "string"
        },
        "reaction": {
          "type": "string",
          "default": "Like"
        }
      }
    },
...
  }
}